+++ pet_website_app/routes/login.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const {User, UserClass} = require ('../models/user');
+const fetchProducts = require ('../controllers/fetchController');
 
 
 router.post('/', async (req, res) => {
@@ -9,10 +10,11 @@ router.post('/', async (req, res) => {
     try {
     const ret = await User.find({userName: username, password: password});
     const user = ret[0];
+    productObjects = await fetchProducts();
     if(ret.length != 0 && !user.isAdmin) {
-        res.render('pages/products', {user: ret, title: "Products"});
+        res.render('pages/products', {user: ret, title: "Products", products: productObjects});
     }else if(ret.length != 0 && user.isAdmin){
-        res.render('pages/adminPage', {title: 'Admin'});
+        res.render('pages/adminPage', {title: 'Admin', products: productObjects});
     } else res.redirect('/');
     } catch(error){error("Error: getting user from database")}
 })
